pyproject.toml
[build-system]
requires = ["setuptools >= 61.0", "setuptools-scm>=8.0"]
build-backend = "setuptools.build_meta"

[project]
name = "ez-vc"
version = "1.1.8"
description = "EZ-VC based off F5-TTS: A Fairytaler that Fakes Fluent and Faithful Speech with Flow Matching"
readme = "README.md"
license = {text = "MIT License"}
classifiers = [
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
]
requires-python = ">=3.10,<3.13"
dependencies = [
    # "accelerate>=0.33.0",
    # "bitsandbytes>0.37.0; platform_machine != 'arm64' and platform_system != 'Darwin'",
    "cached_path",
    # "click",
    # "datasets",
    # "ema_pytorch>=0.5.2",
    # "gradio>=5.0.0",
    "hydra-core>=1.3.0",
    "jieba",
    "librosa",
    "matplotlib",
    "numpy<=1.26.4",
    "pydantic<=2.10.6",
    "pydub",
    "pypinyin",
    "safetensors",
    "soundfile",
    # "tomli",
    "torch>=2.0.0,<2.9.0",
    "torchaudio>=2.0.0,<2.9.0",
    "torchdiffeq",
    "tqdm>=4.65.0",
    "transformers<=4.57.3",
    # "transformers_stream_generator",
    # "unidecode",
    "vocos",
    # "wandb",
    "x_transformers>=1.31.14",
    "espnet @ git+https://github.com/wanchichen/espnet.git@ssl",
    "soxr>=1.1.0",
    "ipykernel",
    "bigvgan",
]

[project.optional-dependencies]
eval = [
    "faster_whisper>=0.10.1",
    "funasr",
    "jiwer",
    "modelscope",
    "zhconv",
    "zhon",
]

[project.urls]
Homepage = "https://github.com/SWivid/F5-TTS"

[tool.setuptools.packages.find]
where = ["src"]
include = ["f5_tts*"]

[tool.setuptools.package-data]
"x_vc.default" = ["*.yaml", "*.json"]
